fix: do not hydrate single root elements
e.g. I have a component with several v-if v-else-ifs and some other logic in there, but at the end it only exposes one single root element. due to the check of childElementCount this was always hydrated, even though it shouldn't. Checking for nodeType seems more reliable and is fully supported in all browsers.